Refactors constants dir
