+++ src/app/logger.ts
@@ -1,6 +1,6 @@
 import type { Server } from "@modelcontextprotocol/sdk/server";
 
-import { APP_NAME, DEFAULT_LOG_LEVEL, LOG_LEVEL, VALID_LOG_LEVELS } from "../constants.ts";
+import { APP_NAME, DEFAULT_LOG_LEVEL, LOG_LEVEL, VALID_LOG_LEVELS } from "$/constants";
 import { SetLevelRequestSchema } from "../schemas.ts";
 import type { LogData, LogLevelKey, LogLevelValue } from "../types.ts";
