+++ src/constants/env.ts
@@ -1,0 +1,10 @@
+export const ENV_VARS = {
+  PORT: "MCP_PORT",
+  HOSTNAME: "MCP_HOSTNAME",
+  LOG: "MCP_LOG_LEVEL",
+  ALLOWED_ORIGINS: "MCP_ALLOWED_ORIGINS",
+  ALLOWED_HOSTS: "MCP_ALLOWED_HOSTS",
+  HEADERS: "MCP_HEADERS",
+  NO_HTTP: "MCP_NO_HTTP",
+  NO_STDIO: "MCP_NO_STDIO",
+} as const;
